Implement `asset:put` endpoint for asset upload
Ref. #107
